Use a transposition table to avoid re-searching common sub-trees.
At this table size, one of my more complex sample instances does a
complete search in half the time of a 4096 entry table, and finds an
optimal solution in 5 minutes instead of 10. I don't have the patience
to compare it to a search without the transposition table.

A simpler instance ran in 21.77 seconds without the table, and 12.8
seconds with. The optimal solution was found in 6 seconds without, and 4
seconds with.